refactor(deps): expose NewPrivateClient function
